src/connectors/websocket_connector.py
import websocket

def connect_websocket(url: str):
    """Optional WebSocket connection (not required for basic extraction)."""
    try:
        ws = websocket.WebSocket()
        ws.connect(url)
        print("Connected to WebSocket")
        ws.close()
    except Exception as e:
        print(f"WebSocket connection failed: {e}")
