requirements.txt
playwright>=1.48.0
pydantic>=2.9.0
pyyaml>=6.0
requests>=2.32.0
aiohttp>=3.9.0
websocket-client
PyYAML
httpx[http2]

